backend/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
flask-sqlalchemy==3.0.3
psycopg2-binary==2.9.5
PyJWT==2.6.0
Werkzeug==2.2.3
gunicorn==20.1.0
python-dotenv==1.0.0